refactor: pkg adjustment
